requirements.txt
# Arayüz ve Veri İşleme
streamlit==1.55.0
pandas==2.3.3
numpy==2.4.3

# Yapay Zeka ve RAG Altyapısı (LlamaIndex & LangChain)
llama-index==0.12.18
llama-index-llms-groq==0.5.0
langchain==1.2.14
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-huggingface==1.2.1
langchain-qdrant==1.1.0
langchain-groq==1.1.2

# Vektör Veritabanı ve Embedding
qdrant-client==1.17.1
sentence-transformers==5.3.0

# Derin Öğrenme ve Model Destek (Hata almamak için şart)
torch==2.11.0
transformers==4.57.6
scipy==1.17.1
scikit-learn==1.8.0

# Yardımcı Araçlar
pypdf==6.9.2
python-dotenv==1.2.2
requests==2.33.1
tenacity==9.1.4
